all test cases passed

+++ app/Http/Controllers/MonsterController.php
@@ -43,10 +43,29 @@ class MonsterController extends Controller
 
         return response()->json(
             [
-                'data' => $this->$monsters
+                'data' => $monsters
             ],
             Response::HTTP_OK
         );
+    }
+
+    /**
+     * Retrieve a monster by its ID.
+     *
+     * @param int $id
+     * @return JsonResponse
+     */
+    public function getMonster(int $id): JsonResponse
+    {
+        $monster = Monster::find($id);
+
+        if ($monster) {
+            return response()->json(['data' => $monster], Response::HTTP_OK);
+
+        }
+        return response()->json(['error' => 'The monster does not exists.'], Response::HTTP_NOT_FOUND);
+
+
     }
     /**
      * Create new monster.
@@ -86,18 +105,24 @@ class MonsterController extends Controller
     public function update(Request $request): JsonResponse
     {
         $monsterId = $request->route('id');
-        $newMonster = $request->only([
-            'name',
-            'attack',
-            'defense',
-            'hp',
-            'speed',
-            'imageUrl'
-        ]);
+        $monster = Monster::find($monsterId);
+        if($monster){
 
-        $result = $this->monsterService->getMonsterById($monsterId);
-        $this->monsterService->updateMonster($monsterId, $newMonster);
-        return response()->json('', Response::HTTP_OK);
+            $newMonster = $request->only([
+                'name',
+                'attack',
+                'defense',
+                'hp',
+                'speed',
+                'imageUrl'
+            ]);
+            $this->monsterService->updateMonster($monsterId, $newMonster);
+            return response()->json('', Response::HTTP_OK);
+
+        }
+
+        return response()->json(['message'=>'The monster does not exists.'], Response::HTTP_NOT_FOUND);
+
     }
 
     /**
@@ -112,8 +137,13 @@ class MonsterController extends Controller
     {
         $monsterId = $request->route('id');
         $result = $this->monsterService->getMonsterById($monsterId);
-        $this->monsterService->removeMonster($monsterId);
-        return response()->json('', Response::HTTP_NO_CONTENT);
+        if($result){
+            $this->monsterService->removeMonster($monsterId);
+            return response()->json('', Response::HTTP_NO_CONTENT);
+        }
+        return response()->json(['message'=>'The monster does not exists.'], Response::HTTP_NOT_FOUND);
+
+
     }
 
     public function importCsv(Request $request): JsonResponse
@@ -134,7 +164,7 @@ class MonsterController extends Controller
 
                 try {
                     $this->monsterService->importMonster($rowData, $csv_data);
-                    return response()->json(['data' => 'Records were imported successfully.'], Response::HTTP_OK);
+                    return response()->json(['message' => 'Records were imported successfully.'], Response::HTTP_OK);
                 } catch (QueryException $e) {
                     return response()->json(['message' => 'Incomplete data, check your file.'], Response::HTTP_BAD_REQUEST);
                 } catch (Exception $e) {
